Added bash script 'download_size.sh'
This script tries to obtain the size of the download file when the URL
to the file is passed as an argument to the script. It uses 'wget' and
'curl' to obtain the information. The script outputs 3 values as it
caluclates the size using 3 ways. For more information, read the script.