Added script to display process network activity
The script 'process_network_monitor.sh' tries to display the network
activity and statistics of a process that matches the PID or search term
that is given as an argument to the script. The script uses the 'nettop'
package to display the process's network activity. For more information,
read the script.