Ability to run 'nettop' in superuser mode
Have added the ability to run 'nettop' in superuser mode using 'sudo',
if needed. The script first checks which user the process is running
as. If the process is running as 'root', the user is informed about it
and some information about the process is printed on screen. Then, the
script asks if the user would like to run 'nettop' with 'sudo' in order
to view the network activity of that process. If the user enters 'y' or
'Y', the script will prompt the user for the administrator password. If
any other character is entered, the script will exit.