Added new script 'stop_swarm_nodes.sh'
This script uses the output of the 'docker-machine ls' command and
filters out the names of the machines which are running and adds them to
an array. Then, it cycles through each element of the array(which are
the names of the individual machines) and runs the 'docker-machine stop'
command against them.