Ability to accept arguments with expanded scope
Have added the ability to accept three command line arguments to the
script. Currently, it accepts 'start', 'stop', and 'help'. The 'start'
argument, as the name states, starts all the machines that are in a
stopped state using 'docker-machine'. The 'stop' argument stops all the
machines that are in a running state using the 'docker-machine' command.
The 'help' argument displays a usage message that details how the script
can be used and what arguments it accepts. '-h', '--help', 'help', and
any other argument that doesn't match any of the aforementioned
arguments will display the same usage message.