Fixed error in 'ARRAY'
The bash commands to be run whose output is to be fed to the array named
'ARRAY' in the 'start' case had an error where the 'grep' command
filtered out the machines in the 'Running' state whereas it needs to
filter out the machines that are in 'Stopped' state. This has been
fixed. That is, 'Running' has been changed to 'Stopped'.