Improved 'rpi_sdcard_imager.sh' script
Have added the capability to find and extract files with the extensions
'.xz', '.bz2', '.7z'. Previously, temporary files that were created
for storage of useful content were created in the directory the script
was run from. This has been changed to ensure that the 'mktemp' command
creates the temporary files in the directory given in the $TMPDIR
variable. So far, this has been tested on the local OSX machine. It is
yet to be seen if all parts of the script can function with a Linux
toolset.