Added scripts to check pulls on Docker Hub
Have added two scripts that check the number of pulls for a Docker image
whose name is returned by a search certain search result.

One of the scripts, 'docker_hub_pull_checker.sh' checks to see the
number of pulls on the repository returned by the search term
'harshavardhanj'.

The other script, 'docker_hub_pull_checker.sh' checks to see the number
of pulls on the repository returned by the search term that is given as
an argument to the script.