New script 'docker_hub_image_stat.sh'
Have added new script 'docker_hub_image_stat.sh'. This script takes one
argument and searches DockerHub for the text passed in that argument. It
displays five of the repositories and/or images, the number of stars
they have, and the number of total pulls of each image.

For now, only formatting needs to be fixed. When the names of some of
the repositories are fairly long, the number of stars and number of
pulls are being offset due to which they don't line up with the header
printed previously.