Fixed 'printf' formatting
Have fixed the formatting issues. The results are all aligned with the
header of the output now. The only issue that is left to be resolved is
the replacement of the forward slash with '\u002F'. Regardless of
whether 'cURL' or 'wget' is used, the dowloaded HTML file has all
forward slashes replaced by '\u002F'. This shows up in the output of the
results too. Need to find out why this happens.