Fixed appearance of '\u002F' instead of '/'
Have fixed the appearance of the string '\u002F' instead of '/'. Used
'sed' to replace the string '\u002F' with '/' in the array that has the
repository names stored. Currently, the output of the script looks as
follows. For this example, the script is run with the 'postgres'
argument

    $ ./docker_hub_image_stat.sh postgres

The output:

REPO                                      STARS           PULLS
---------------------------------------------------------------
postgres                                   5077       368357956
circleci/postgres                             6        19419228
circleci/postgres                            45         2585643
orchardup/postgresql                         19        16899122
orchardup/postgresql                         41         2668025