Major changes to 'rpi_sdcard_imager.sh'
Have made multiple major changes to the 'rpi_sdcard_imager.sh' script
file. Added the option to configure the SD card to enable the SBC to be
able to connect to a certain WiFi network on boot, and to allow SSH
access on boot. Have fixed all the errors generated by 'shellcheck'.