Added script 'colorise_output.sh'
Added new script 'colorise_output.sh'. This script prints a string in a
coloured output. It accepts two arguments. The first argument is the
colour that the text is to be printed in. The second argument is the
actual string/text that is to be printed. Note that the string that is
to be printed should be within quotation marks for it be accepted as one
argument. For a list of colours, take a look at the script.