Moved functions. Variable changed to local
Have moved the 'start_vm' function to the beginning of the script. The
functions are listed according to the order in which they are called
i.e.

1). Start VM
2). Ping VM
3). Check if port 22 is accepting connections
4). Check if SSH connections can be made to VM using the 'uptime'
command run on the VM via SSH
5). If the VM responds back with an uptime report, start an SSH session.

---

The value of the IP address of the VM was declared as a global variable
within the 'ping_vm' function. This had been done because this variable
was accessed within the 'ssh_port' function as well. Now, the IP address
has been declared as a local(read-only) variable within the 'ping_vm'
and 'ssh_port' functions.