Added case-senstive matching
The first argument to the script, which is the colour in which the
string(which is the second argument) is to be printed, would previously
match only lowercase values for the colour such as "red" or "Red".

Now, any variation of case matches the string i.e., "Red", "rEd", "RED",
and 'reD' all match the colour 'red'.

Have also changed the variables to local, read-only type.