Added new project : 'usb_flash'
This folder contains the script files pertaining to the 'usb_flash'
script. Each script contains either one function or a set of functions
and/or modules that take care of one 'area' of the project.

For example, the 'general_functions.sh' script contains functions and
modules that will be used by all other scripts. It contains functions
that facilitate logging, printing to standard error, etc. The
'os_check.sh' script contains function(s) that verify the type of OS the
script is running on. Each such function has well-defined input and
output. Any deviation will cause them to exit with a return code of 1.

The functions and modules are structured in such a way that they accept
either one or more input values and return an output value corresponding
to the input. This way, the function can be modified in any way as long
as the output values are maintained. This allows for modularity. Also,
the underlying working of each function can be abstracted away. This
way, for example, if some change needs to be made to the function that
prints to a log file, the changes can be made as long as the final
result, which is printing to a log file, is the same. This way, the
functions can be utilised by other scripts without worrying about how
they work, as long as the inputs are defined correctly.

This is a work in progress. Multiple changes will be made to each
file/script. The scripts that have '(Stable)' in their version number
work as intended and have been tested.