New functions added to 'general_functions.sh' file
Have added new functions. The most important one is the 'import_files'
function. This function can be called if a script needs to be imported.
This function has some checks built-in to ensure that the same file is
not imported muliple times. Sometimes, when one script has an import
statement pointing to another file, and when that file has an import
statement pointing to the first file, they can enter into an infinite
loop by continuously importing the other function. The 'import_files'
function uses an 'import status' variable(a global variable) to ensure
that a file, once imported, is not imported again.

This is done by generating a unique 'import status' variable for each
script name(or file name), and declaring a global varible with that
name. When a file is imported, the 'import status' variable is declared
and its value is set to a certain value. In the process of importing
files by using the 'import_files' function, the function checks to see
if the 'import status' variable corresponding to the file name has been
set to the certain value. If the variable has not been set, the file is
imported. If the variable has been set to a certain value(which is 1),
the file is not imported. This way, a file will not be imported more
than once even if the 'import_status' function is called multiple times
to import a certain file.

There are other functions as well. One is the 'log_to_file' function
which helps create a log file, and print strings to it. If the log file
has not been created, the string will not be appended.