Major changes to 'var_file_collection.sh'
Have made some major changes to the 'var_file_collection.sh' script
file. The file was starting to resemble a monolithic piece of code
inspite of the attempt to break it up into functions. Now, a lot of code
that was being repeated has been incorporated into functions. Currently,
the function/script does not work. There are many more changes to be
made to the script. The function 'var_file_collection__vars_files_array'
has been changed quite a bit. The names of the arrays are no longer
directly referenced throughout the code. Instead, they have been
replaced by a nameref variable. This way, the name of the arrays can be
changed if necessary without affecting the rest of the code base.