Added new feature to 'colourise_output.sh' script.
Added a new feature to the 'colourise_output.sh' script. The script can
now print coloured text to standard error if needed. The user will need
to set the '--error' or '-e' flag to '1' in order to send the text to
standard error instead of standard output which is where text will be
printed by default. Setting '--error' or '-e' to 0 is the same as the
default option where text is printed to standard output.

This file has also been added to the 'usb_flash' directory as its
functions will be used by other scripts/files within this directory.