Added new feature to 'colourise_output.sh' script
Added a new feature to the 'colourise_output.sh' script where the
printing of a newline after every function call can be overridden if
needed. This was done to provide flexibility of usage of the function.
Previously, in an older version of the script, the behaviour was to
print all the text on the same line. Even if the function was called for
the second time, the text would be printed on the same line. This
required the user to add a `print '\n'` command after every function
call if they wanted a newline to be added.

In later versions of the script, the error feature was added in which
the user could direct the text to be printed to standard error instead
of standard output. The default behaviour was to print to standard
output, but this could be overridden by using the '--error' flag.

In this version, the option to print a newline is provided. By default,
after every function call, a new line is printed. This can be modified
by passing '--newline 0' to the function which will result in no new
line being printed. Therefore, the next function call will print the
text on the same line as before.

'--newline 0' suppresses the newline at the end
'--newline 1' prints a newline at the end, which is the default