Major changes to 'var_file_collection.sh' file
Have made major changes to the 'var_file_collection.sh' script. Two new
functions have been added. The first one is the 'cleanup' submodule
which unsets the global namerefs that point to the arrays that hold the
variables and files that are to be kept track of. To be more precise,
the namerefs 'VarsArrayName' and 'FilesArrayName' point to the arrays
'VarsArray' and 'FilesArray' respectively. During the clean up phase,
apart from the variables in the 'VarsArray' and the files in the
'FilesArray' being unset and deleted, the namerefs that point to these
two arrays will need to be unset as well because these namerefs are
global variables too. Also, in order to utilise this submodule, another
option has been added to the 'vars_files_array' submodule, namely the
'cleanup' option/argument. Now, the user can pass 'cleanup' as an
argument to the 'vars_files_array' submodule and the actions described
above will be taken.

The second function is a 'main' function which is essentially resposible
for sourcing dependencies. Since this script depends on the
'general_functions.sh' file, it will need to be imported/sourced. But,
the 'general_functions.sh' file also imports the 'var_file_collection'
script. This results in recursive importing which can continue
endlessly. To prevent this, the 'import_files' function has been written
in the 'general_functions.sh' script. This function imports a script
that is passed to it, and then sets a global variable that is unique as
it is named after the file that is being imported. Then, supposing that
the same file is being passed to the 'import_files' function, it first
checks if the unique global variable that corresponds to the script has
been set, and if it has, the file is not imported. This way, infinite
recursive importing is prevented. Now, it is possible to just add an
`source ./general_functions.sh` and `import_files
./general_functions.sh` statement at the end of the script. Instead, it
is easier to include all these statements inside a 'main' function along
with any other commands to be executed during the initial stages. This
function can then be called. The reason this is done is because the
files to be imported/sourced are usually stored in an array because it
becomes easy to pass the list of files to the 'import_status' function
by running `import_status "${importArray[@]}"`. But, almost each file
will need to import dependencies, if the arrays within those files are
declared outside a function, they have a global scope. So, one will need
to have unique names for this array in each of those files so that they
don't conflict with each other. This becomes tedious and hard to
maintain as the number of modules/dependencies increase. Instead, it is
easier to create a 'main' function within each script and declare a
local array within that function. Now, this array does not need to have
a unique name as its scope is local and there is no danger of it
conflicting with other array names from other files.

Finally, most of the functions within the 'var_file_collection.sh' file
had some bugs and issues which have been resolved. Bugs will be fixed as
and when they are encountered during testing or usage. The printing of
error messages and statuses has been made prettier and more robust than
before.