Fixed bugs. Added new features.
Added new features/functions to the 'general_functions.sh' script. The
first feature is the colour-coded printing of error messages in the
'print_err' function. The function now provides the option to print the
function name which is calling it, and the line number on which the
function call is being made. This way, error messages are easily
traceable to the right function and the line number. The 'print_err'
function utilises the 'colourise' function from the
'colourise_output.sh' script for this.

The second "feature" involves the 'log_to_file' function which was
lacking in many areas. Previously the function could just initialise a
log file when a user provided `init /path/to/file` as an argument. If
the file already existed or if the global variable that pointed to that
file was already set, the function didn't handle it well. Now, sane and
descriptive error messages are printed and wherever possible, the
appropriate action is taken.

Finally, just like in the 'var_file_collection.sh' script, a main
function has been added to this script. This is to aid in the importing
of other dependencies. The issues that this solved are described in the
previous commit involving the 'var_file_collection.sh' script.