Added new feature to 'cleanup.sh' file
The new "feature" added to the 'cleanup.sh' file is the function call to
the 'vars_files_array' function with the 'cleanup' argument which unsets
the global namerefs 'VarsArrayName' and 'FilesArrayName'. Also, pretty
printing of error messages has been enabled. Finally, a main function
has been added which, when called, takes care of importing the
dependencies.