Some changes to the 'check_image_file.sh' file
Have made some improvements to the 'check_image_file.sh' script. One of
which is the ability to now pretty print error messages which helps zone
in on the source of the error. Secondly, a main function has been added
which takes care of the importing of dependencies. Also, there seems to
be a case where this script does not work as it should. Will look into
it soon and fix it.